requirements.txt
django==1.6.6
djangorestframework
django-bootstrap3